only delete last batch when theres more than 1 batch

+++ app/models/batch.rb
@@ -31,6 +31,7 @@ class Batch < ActiveRecord::Base
   end
 
   def destroy_previous_batch
+    return unless Batch.count > 2
     Batch.first.destroy
   end
